+++ files.go
@@ -5,6 +5,50 @@ import (
 	"strings"
 )
 
+func fileStartsWithString(path string, str string) bool {
+	f, err := os.Open(path)
+	if err != nil {
+		return false
+	}
+	defer f.Close()
+
+	buff := make([]byte, len(str))
+	_, readErr := f.Read(buff)
+	if readErr != nil {
+		return false
+	}
+
+	return str == string(buff)
+}
+
+func fileEndsWithString(path string, info os.FileInfo, str string) bool {
+	f, err := os.Open(path)
+	if err != nil {
+		return false
+	}
+	defer f.Close()
+
+	bsize := len(str)
+	buff := make([]byte, bsize)
+	cbuff := make([]byte, 1)
+
+	j := bsize - 1
+	for i := info.Size() - 1; i >= 0 && j >= 0; i-- {
+		_, readErr := f.ReadAt(cbuff, i)
+		if readErr != nil {
+			return false
+		}
+
+		if cbuff[0] >= 32 || j < bsize - 1 {
+			buff[j] = cbuff[0]
+			j--
+		}
+	}
+
+	return str == string(buff)
+
+}
+
 func fileContainsString(path string, str string) bool {
 	bsize := len(str)
 	f, err := os.Open(path)