+++ RELEASENOTES.md
@@ -9,6 +9,7 @@
 * Renamed "contents" keyword to "content".
 * Updated "size" display and added units suffixes to integer values.
 * Added support for "name" and "path" string equality checking.
+* Added support for "content" to "startswith" and "endswith" queries.
 
 ## 0.0.1